+++ user-commands/changelog.md
@@ -6,7 +6,7 @@
 
 **Example:**
 
-<span style="color:#00FFFF"><b>TRIODEOfficial</b></span>: !command add hug @{{displayname}} hugs @{{message}}!<br/>
+<span style="color:#00FFFF"><b>TRIODEOfficial</b></span>: !command add hug @{displayname} hugs @{message}!<br/>
 <span style="color:#5555FF"><b>LumiaStreamBot</b></span>: !hug has been created.<br/>
 <span style="color:#FF69B4"><b>Lexie</b></span>: !hug @TRIODEOfficial</span><br/>
 <span style="color:#5555FF"><b>LumiaStreamBot</b></span>: <span style="color:#FF69B4">@Lexie</span> hugs <span style="color:#00FFFF">@TRIODEOfficial</span>!<br/>
@@ -17,7 +17,7 @@
 ### v0.9.0-beta.1
 - Initial beta release
 - Add, edit, delete, and list custom commands
-- Support for {{displayname}} and {{message}} variables
+- Support for {displayname} and {message} variables
 
 ---
 Made with ❤️ in San Francisco, California by [@TRIODEOfficial](https://twitch.tv/TRIODEOfficial)